Milestone_4/airflow_milestone4/requirements.txt
scikit-learn
dash
pyarrow<10.1.0,>=10.0.1
pandas
numpy 
seaborn
matplotlib
googlemaps
SQLAlchemy
psycopg2-binary
plotly